requirements.txt
azure-functions
azurefunctions-extensions-http-fastapi
openai